+++ desktop/CHANGELOG.md
@@ -1,19 +1,43 @@
 # Changelog | Ahau desktop
 
+## v2.14.0 (Draft)
+
+### Features
+
+- tree search bar
+  - uncollapses all nodes from the root/focus node to the searched profile
+  - highlights the path
+
+## Bug Fixes
+
+- N/A
+
+### Improvements
+
+- show timeline button on mobile
+Persons list
+  - improve ui for mobile
+  - update search bar and title to be consistent with existing styling
+WhakapapaIndex
+  - update the whakapapa cards to be the same as the collection cards when on mobile
 
 ## v2.13.0
+
 ### Features
+
 * added images and multiple filter and toggle columns on peoples list
 * removed the background color on the whakapapa icons
 * added a loader whakapapa show when loading profiles and connections
 * added a loader in place of the node menu on a person when a whakapapa is still loading. This limits what a user can do when a graph is still loading and will prevent the graph trying to make changes while loading
 
-### Fixes
+## Bug Fixes
+
 * removed the toggle timeline/archive button when there are no stories
 * fixed a csv import bug where the where it was importing properly if a person was in the csv more than once
 * changed how we get and load the node profiles in a graph to speed up whakapapa loading
 
-### Changes
+### Improvements
+
 * disabled the scroll on peoples list
 * moved table filters to a lib so i could use the filter functionality in the peoples list
 * added all profiles to store to enable faster loading of whakapapas